% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{anova_lm}
\alias{anova_lm}
\title{Obtain ANOVA statistics for a linear model.}
\usage{
anova_lm(lmobject)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function.}
}
\value{
A dataframe with various labelled ANOVA statistics.
}
\description{
Obtain ANOVA statistics for a linear model.
}
