% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{coefficient_confidence_lm}
\alias{coefficient_confidence_lm}
\title{(Stat 5100) Calculate confidence intervals for the coefficients of a linear
model. This function pulls info from the lm object and constructs a
confidence interval of the form "coefficient est +- critical value * std error"}
\usage{
coefficient_confidence_lm(lmobject, confidence = 0.95, simul = FALSE)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function}

\item{confidence}{A level of confidence for the interval expressed as a proportion}
}
\value{
A coefficient matrix for the model that contains lower and upper
confidence interval values for the coefficient estimate.
}
\description{
(Stat 5100) Calculate confidence intervals for the coefficients of a linear
model. This function pulls info from the lm object and constructs a
confidence interval of the form "coefficient est +- critical value * std error"
}
