% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{cor_normality_lm}
\alias{cor_normality_lm}
\title{(Stat 5100) Correlation test of normality function. This function will give
output on the correlation between the residuals and the expected residuals
under a normal distribution. Table B.6 from the notes will give you what
correlation numeric values would reject/fail to reject a null hypothesis
that the residuals have a normal distribution.}
\usage{
cor_normality_lm(lmobject)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function}
}
\value{
Console output with a correlation matrix of the residuals and
expected residuals
}
\description{
(Stat 5100) Correlation test of normality function. This function will give
output on the correlation between the residuals and the expected residuals
under a normal distribution. Table B.6 from the notes will give you what
correlation numeric values would reject/fail to reject a null hypothesis
that the residuals have a normal distribution.
}
