% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{fit_plot}
\alias{fit_plot}
\title{(Stat 5100 function) Plot a scatterplot of the original data with a
regression line overlaid on top. This can help you decide on how well a
regression line fits the data.}
\usage{
fit_plot(lmobject, ...)
}
\arguments{
\item{lmobject}{An object of type 'lm' created from the lm() function.}
}
\value{
A plot object of the regression model and the original data.
}
\description{
(Stat 5100 function) Plot a scatterplot of the original data with a
regression line overlaid on top. This can help you decide on how well a
regression line fits the data.
}
