% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ftest_lackfit_lm}
\alias{ftest_lackfit_lm}
\title{(Stat 5100) F-test for lack of fit. A test for determining if too much
of the error in prediction is due to the lack of model fit.}
\usage{
ftest_lackfit_lm(lmobject)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function.}
}
\value{
An ANOVA result that gives an F-statistic and p-value for this
test.
}
\description{
(Stat 5100) F-test for lack of fit. A test for determining if too much
of the error in prediction is due to the lack of model fit.
}
