% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{qq_plot}
\alias{qq_plot}
\title{(Stat 5100 function) Obtain a QQ plot for a linear model. This plot can
help you assess the normality assumption of a linear regression model. This
plot will plot theoretical quantiles of a normal distribution against the
residuals.}
\usage{
qq_plot(lmobject)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function.}
}
\value{
A QQ plot object.
}
\description{
(Stat 5100 function) Obtain a QQ plot for a linear model. This plot can
help you assess the normality assumption of a linear regression model. This
plot will plot theoretical quantiles of a normal distribution against the
residuals.
}
