% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/library.R
\name{residual_hist}
\alias{residual_hist}
\title{(Stat 5100 function) Obtain a histogram of residuals for a linear model.
This function also plots a normal curve on top of the histogram with the
same mean and standard deviation. This helps you assess the validity of the
normality assumption of residuals in linear regression.}
\usage{
residual_hist(lmobject)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function.}
}
\value{
A histogram of the residuals along with a theoretical normal curve
with the same mean and standard deviation.
}
\description{
(Stat 5100 function) Obtain a histogram of residuals for a linear model.
This function also plots a normal curve on top of the histogram with the
same mean and standard deviation. This helps you assess the validity of the
normality assumption of residuals in linear regression.
}
