% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{residual_plot}
\alias{residual_plot}
\title{(Stat 5100 function) Obtain a residual plot for a linear model. This helps you decide on the
appropriateness of linear regression model assumptions.}
\usage{
residual_plot(lmobject, ...)
}
\arguments{
\item{lmobject}{An object of type 'lm' created from the lm() function.}
}
\value{
A plot object of the residuals plotted against the fitted values.
}
\description{
(Stat 5100 function) Obtain a residual plot for a linear model. This helps you decide on the
appropriateness of linear regression model assumptions.
}
