% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{seq_plot}
\alias{seq_plot}
\title{(Stat 5100 function) Obtain a sequence plot of residuals. This is a plot
that plots the order of the residuals in the data agains the residuals
themselves. This can help you identify a possible violation of
independence in the residuals.}
\usage{
seq_plot(lmobject, ...)
}
\arguments{
\item{lmobject}{A linear model object from the lm() function.}
}
\value{
A plot object with the residuals plotted in order.
}
\description{
(Stat 5100 function) Obtain a sequence plot of residuals. This is a plot
that plots the order of the residuals in the data agains the residuals
themselves. This can help you identify a possible violation of
independence in the residuals.
}
