% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simul_mean_prediction_limits}
\alias{simul_mean_prediction_limits}
\title{(Stat 5100) confidence interval for the mean response variable at a given
X-level.}
\usage{
simul_mean_prediction_limits(lmobject, newdata, confidence = 0.95)
}
\arguments{
\item{lmobject}{The linear model object from the lm() function}

\item{newdata}{A dataframe with columns named the same thing as the predictor variables
used to create the model.}

\item{confidence}{A number between 0 and 1 that indicates the confidence level}
}
\description{
(Stat 5100) confidence interval for the mean response variable at a given
X-level.
}
