% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{simul_prediction_limits}
\alias{simul_prediction_limits}
\title{(Stat 5100) Get a confidence interval for an individual new prediction.
Note that this is different from the confidence interval for the mean of
the response variable at a particular X-profile.}
\usage{
simul_prediction_limits(lmobject, newdata, confidence = 0.95)
}
\arguments{
\item{lmobject}{The linear model object from the lm() function.}

\item{newdata}{A dataframe that contains the same variables used to train
the lm object (no response variable needed).}

\item{confidence}{The simultaneous confidence level for the predictions.}
}
\description{
(Stat 5100) Get a confidence interval for an individual new prediction.
Note that this is different from the confidence interval for the mean of
the response variable at a particular X-profile.
}
