% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{visual_assumptions}
\alias{visual_assumptions}
\title{(Stat 5100 function) A combination of the 4 main graphical checks to
assess whether model assumptions are satisfied or not. This is a way
to clean up the code and make it a bit easier.}
\usage{
visual_assumptions(lmobject, ...)
}
\arguments{
\item{lmobject}{The linear model object from the lm() function}
}
\value{
A 2x2 plot with some main graphical checks we use.
}
\description{
(Stat 5100 function) A combination of the 4 main graphical checks to
assess whether model assumptions are satisfied or not. This is a way
to clean up the code and make it a bit easier.
}
